Use functions in `MLIExec` (#13)
* Clean up helper functions, formatting for messages

* Permit multiple return values

* Use print for legacy APFloat

* Fix breakage in parsing due to whitespace

+++ test/llvm_ir/trunc/test_trunc.mlir
@@ -1,4 +1,4 @@
-func.func @main(%arg0: i64) -> i32 {
-  %res = llvm.trunc %arg0 : i64 to i32
-  llvm.return %res : i32
+func.func @main(%arg0: i32) -> i8 {
+  %res = llvm.trunc %arg0 : i32 to i8
+  llvm.return %res : i8
 }
